Reorganized code to seperate folders
