+++ calculation/inc/calculation/ICalculator.hpp
