+++ calculation/inc/calculation/ampPhaseCalc.hpp
